fix: address review findings across the pod-list and inspect work
Three reviews over the branch turned up 2 bugs, 28 risks and 8 nits. This
clears all of them.

Two real bugs
-------------
* Cross-namespace pod contamination. The pod delta stream is cluster-wide
  while the backing list is namespaced, and `acceptsPodDelta` matched on
  labels alone - so a pod from another namespace with the same labels joined
  the list after mount. `app=web` in both staging and production is the common
  case, and workload panels hide the namespace column, so it was invisible.
  `acceptsPodDelta` now takes the namespace and checks it first; the Node
  panel passes null because its pods legitimately span namespaces.
* EventsTab skipped its INITIAL fetch when the window was backgrounded, not
  just its poll, leaving the tab stuck on "Loading events..." until it was
  visible AND a 10s tick landed. Gated the interval only, matching
  DetailPanel.

Pod list effects
----------------
* `unsubscribe_resource` decrements a SHARED refcount and was called even when
  the subscribe never ran - stealing the Pods table's watcher, deterministically
  under StrictMode. Now guarded on having actually subscribed.
* `knownRef` kept the previous subject's uids across a subject change, so the
  old subject's pods passed the known-uid short-circuit and became permanent.
* Both effects share one map, so the identity guard couldn't separate them. A
  monotonic fetch sequence now means only the newest-issued fetch may write.
* No in-flight guard: `detailVersion` bumps per debounced watcher delta, so a
  churning rollout fired several LISTs a second. Coalesced to one.
* A refetch dropped pods the delta stream had admitted after it was issued.
  Those uids are now tracked and merged back.

Inspect drawer
--------------
* PodsTab had `refetchKey={0}`, so its list froze at mount - and a
  matchExpressions selector can't admit pods from the stream, so a scale-up
  never appeared. Now polls.
* Owner chips came from re-matching selectors, which misattributes overlapping
  ones and returns nothing for matchExpressions. The fetch already knows which
  controller returned each pod; that mapping is used instead.
* Events no longer waits on the manifest fetch it doesn't consume, and a
  partial failure now offers a retry rather than only an all-failed one.
* Namespace-qualified labels when a selection spans namespaces - comparing
  `frontend` in two namespaces rendered two identical chips.
* Plurals via a tested helper: "Ingresses", not "Ingresss".
* Loading indicators: LoadingLine's centred layout is height:100%, so the
  padded auto-height wrappers collapsed it into a bar jammed at the top. It
  now goes straight into a column-flex pane, as LogPanel does.

fieldFlatten
------------
* Dotted keys collided with nesting: `{a:{b:1}}` and `{"a.b":2}` both wanted
  path `a.b`, so one silently vanished from the comparison. Real Kubernetes
  keys (`app.kubernetes.io/name`) hit this constantly. Non-plain segments are
  now bracket-quoted.
* Arrays compared positionally, so two identical objects whose containers were
  ordered differently reported every element as changed. Entries now key on
  their Kubernetes merge key (`name`, `port`, `mountPath`, ...) where they have
  one - the server's own rule - and fall back to the index for order-significant
  lists like `args`.

Backend
-------
* `selector_query` silently dropped an unrecognised matchExpressions operator,
  which WIDENS the query - a dropped NotIn means matching pods the workload
  does not own. Pre-existing, but this branch routed port-forward traffic and
  the pod list through it. It now refuses the selector.
* `list_pods_for_workload` was unbounded; capped at MAX_RESOLVED_PODS like the
  log resolver, and its doc no longer claims label-selector matching is
  ownership.
* Job's pod count used i32 arithmetic on apiserver values - debug panics on
  overflow, release wraps negative, inside a projection that must be total.

Also: an Esc while a drawer is open no longer wipes the selection the drawer
was built from (a new `drawerOpen` hotkey layer; the bug also affected
ComparePanel and LogPanel on main), the integration test's isolation assertion
is no longer vacuous, and its namespace waits out a previous run's teardown.

Tests: +30, covering every fix above - cross-namespace admission, backgrounded
load, subscribe/unsubscribe pairing, refetch coalescing, merge-key ordering,
dotted-key quoting, selector operator refusal, and a first PodsTab suite.

+++ crates/kube-ext/src/kinds/jobs.rs
@@ -77,15 +77,18 @@ impl KindSpec for JobSpec {
         json!({
             "namespace": meta.namespace.clone().unwrap_or_default(),
             "name": meta.name.clone().unwrap_or_default(),
-            // Pods *created*, matching `kubectl describe job`'s "Pods Statuses"
-            // line — not a live count. Succeeded pods stay counted here after
-            // the apiserver has garbage-collected them.
+            // `active + succeeded + failed`, matching `kubectl describe job`'s
+            // "Pods Statuses" line. Not a live count: succeeded pods stay
+            // counted after the apiserver garbage-collects them, and the
+            // number dips while pods terminate. Widened to i64 because a
+            // projection must be total — i32 addition on apiserver-supplied
+            // values panics in debug and wraps negative in release.
             //
             // Earns a column where Deployment's didn't: `completions` counts
             // only successes, so a job that failed twice before succeeding
             // reads 1/1 there and 3 here. Deployment/StatefulSet dropped
             // theirs — `ready` already carried the same number.
-            "pods": active + succeeded + failed,
+            "pods": i64::from(active) + i64::from(succeeded) + i64::from(failed),
             "completions": format!("{succeeded}/{desired}"),
             "phase": phase,
             "creation_timestamp": meta.creation_timestamp.as_ref().map(|t| t.0.to_string()),
